+++ R/RcppExports.R
@@ -17,6 +17,7 @@
 #' @param dirichlet not yet implemented
 #' @param proportion_model logical value for modeling the proportions data
 #' @param BVS logical value for implementing Bayesian variable selection
+#' @param CMH logical value for conditional MH or Carlin-Chib augmented MH for Bayesian variable selection
 #' @param threads maximum threads used for parallelization. Default is 1
 #' @param gamma_prior one of \code{c("bernoulli", "MRF")}
 #' @param gamma_sampler one of \code{c("mc3", "bandit")}
@@ -31,7 +32,7 @@
 #' @param datX0 a matrix of mandatory variables
 #' @param datProportionConst an array of cluster-specific proportions
 #'
-run_mcmc <- function(nIter, burnin, thin, tick, n, nsamp, ninit, convex, npoint, dirichlet, proportion_model, BVS, threads, gamma_prior, gamma_sampler, eta_prior, eta_sampler, initList, rangeList, hyperparList, datEvent, datTime, datX, datX0, datProportionConst) {
-    .Call(`_GPTCM_run_mcmc`, nIter, burnin, thin, tick, n, nsamp, ninit, convex, npoint, dirichlet, proportion_model, BVS, threads, gamma_prior, gamma_sampler, eta_prior, eta_sampler, initList, rangeList, hyperparList, datEvent, datTime, datX, datX0, datProportionConst)
+run_mcmc <- function(nIter, burnin, thin, tick, n, nsamp, ninit, convex, npoint, dirichlet, proportion_model, BVS, CMH, threads, gamma_prior, gamma_sampler, eta_prior, eta_sampler, initList, rangeList, hyperparList, datEvent, datTime, datX, datX0, datProportionConst) {
+    .Call(`_GPTCM_run_mcmc`, nIter, burnin, thin, tick, n, nsamp, ninit, convex, npoint, dirichlet, proportion_model, BVS, CMH, threads, gamma_prior, gamma_sampler, eta_prior, eta_sampler, initList, rangeList, hyperparList, datEvent, datTime, datX, datX0, datProportionConst)
 }